In work: Re-structured modules and added modules for melting and freezing simulations. Three tests are failing: Stefan problem and freezing
