+++ phaseflow/__init__.py
@@ -1,6 +1,12 @@
-from . import simulation
-from . import phasechange_simulation
-from . import refinement
 from . import backward_difference_formulas
 from . import helpers
 from . import plotting
+from . import refinement
+
+from . import abstract_simulation
+from . import abstract_phasechange_simulation
+from . import abstract_heated_cavity_phasechange_simulation
+
+from . import cavity_freezing_simulation
+from . import cavity_melting_simulation
+
