+++ phaseflow/helpers.py
@@ -29,6 +29,24 @@ class Point(fenics.Point):
             fenics.Point.__init__(self, coordinates[0], coordinates[1], coordinates[2])
 
             
+class SolutionFile(fenics.XDMFFile):
+    """ This class extends `fenics.XDMFFile` with some minor changes for convenience. 
+    
+    Parameters
+    ----------
+    filepath : string
+    """
+    def __init__(self, filepath):
+
+        fenics.XDMFFile.__init__(self, filepath)
+        
+        self.parameters["functions_share_mesh"] = True  # This refers to the component solution functions.
+
+        self.parameters["flush_output"] = True  # This allows us to view the solution while still running.
+        
+        self.path = filepath  # Mimic the file path attribute from a `file` returned by `open` 
+
+
 def mkdir_p(pathstring):
     """ Make a directory if it doesn't exist.
     
@@ -45,6 +63,17 @@ def mkdir_p(pathstring):
     
     path.mkdir(parents = True, exist_ok = True)
 
+    
+def float_in(float_item, float_collection, tolerance = 1.e-8):
+    
+    for item in float_collection:
+    
+        if abs(float_item - item) < tolerance:
+        
+            return True
+    
+    return False
+    
     
 if __name__=="__main__":
